Adds literal, folded, and single-quote scalar dump
Expands the ways that string values can be dumped to yaml.

Use simple scalar when possible, as before.

If simple is not possible, but single-quoted is allowed (ie, no
unprintables or escaped characters), then use that.

If there are linebreaks, and no lines have trailing or leading spaces,
then use folded or literal.  Prefer literal if all lines are below a
reasonable limit.

If neither literal nor folded are possible, then fall back to
double-quoted.

Fixes #66